+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
+## [0.5.0]  2022-01-28
+### Added
+- Select step can include null value
+- Add Asterisk to Select step
+- Add selectAsteriskFrom function to Builder
+- Add Steps RailRoad in doc
+- Add OrderBy step
+
 ## [0.4.9]  2022-01-27
 ### Fixed
 - bug in extending classes STEPs